+++ controllers/series.controller.ts
@@ -1,6 +1,6 @@
 import { Context, Next } from "koa";
 
-import { getSeriesIndexArray } from "../utils/parsers";
+import { getSeriesIndexArray, getSeriesPosts } from "../utils/parsers";
 
 /**
  * @desc Get series array for index page
@@ -11,12 +11,8 @@ import { getSeriesIndexArray } from "../utils/parsers";
 export const getSeriesIndex = async (ctx: Context, next: Next) => {
   try {
     const seriesArray = await getSeriesIndexArray();
-    console.log(seriesArray);
-
     await ctx.render("seriesindex", { seriesArray });
   } catch (error) {
-    console.log(error);
-
     await next();
   }
 };
@@ -30,15 +26,10 @@ export const getSeriesIndex = async (ctx: Context, next: Next) => {
 export const getSeries = async (ctx: Context, next: Next) => {
   const series = ctx.params.series;
 
-  // this route will always be hit even if series is not an actual series
-  // needs to flow to catch { next() } if the series isn't found
-  // so the filter method next will need to throw an error if the series isn't found/empty
-
   try {
-    await ctx.render("series", { series });
+    const blogLinks = await getSeriesPosts(series);
+    await ctx.render("series", { series, blogLinks });
   } catch (error) {
-    console.log(error);
-
     await next();
   }
 };