+++ routes/home.route.ts
@@ -4,7 +4,8 @@ import Router from "@koa/router";
 const homeRouter = new Router<DefaultState, Context>({ prefix: "/" });
 homeRouter.get("/", async (ctx: Context, next: Next) => {
   try {
-    await ctx.render("home");
+    // await ctx.render("home");
+    ctx.redirect("/blog");
   } catch (error) {
     await next();
   }